% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_columnwise_tiles}
\alias{add_columnwise_tiles}
\title{Create columnwise tiles in an empty tiling (from \code{init_tiling}).}
\usage{
add_columnwise_tiles(tiles, dims)
}
\arguments{
\item{tiles}{A tile list (e.g., tiling$tiles)}

\item{dims}{The dimensions of the data matrix}
}
\value{
A tiling where each column is one tile
}
\description{
Create columnwise tiles in an empty tiling (from \code{init_tiling}).
}
