% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_tile}
\alias{add_tile}
\title{Add a new tile to a tiling}
\usage{
add_tile(tiling, tile, target = "tile_list_user", store = TRUE)
}
\arguments{
\item{tiling}{A tiling}

\item{tile}{A new tile to add to the tiling (from \code{create_tile})}

\item{target}{List to which the tile should be added. One of \code{tile_list_user}, \code{tile_list_base}, \code{tile_list_ignore}, or NULL if the tile should not be added anywhere.}

\item{store}{Should the tile be stored in the target}
}
\value{
The tiling with the new tile added
}
\description{
Add a new tile to a tiling
}
