% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bbox}
\alias{bbox}
\title{Bounding box for xy position as lng/lat}
\usage{
bbox(x, y, zoom)
}
\arguments{
\item{x}{x-position on grid}

\item{y}{y-position on grid}

\item{zoom}{zoom level}
}
\value{
list representing bounding box--top left and bottom-right corners.
}
\description{
Bounding box for xy position as lng/lat
}
