% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buffer}
\alias{buffer}
\title{buffer bounding box}
\usage{
buffer(bb, pixels = 1)
}
\arguments{
\item{bb}{bounding box created by `bbox` function}

\item{pixels}{buffer size in pixels}
}
\value{
buffered bounding box
}
\description{
buffer bounding box
}
