% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate the correlation between all pairs of numberic columns in a dataset}
\usage{
calculate_correlation(dataset)
}
\arguments{
\item{dataset}{A dataset (data frame or matrix)}
}
\value{
Correlation matrix
}
\description{
Calculate the correlation between all pairs of numberic columns in a dataset
}
