% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDBcompositeFromModelId.R
\name{CDBcompositeFromModelId}
\alias{CDBcompositeFromModelId}
\title{compositeFromModelId}
\usage{
CDBcompositeFromModelId(
  modelID = NULL,
  ceramic_url = "http://localhost:7007",
  did_private_key = NULL,
  output = "composite.json"
)
}
\arguments{
\item{modelID}{Character.}

\item{ceramic_url}{Character. Ceramic URL}

\item{did_private_key}{Character. Private key that will control composite ??}

\item{output}{name of output file. Default "composite.json"}
}
\value{
Side effect
}
\description{
CDBcompositeFromModelId
}
\examples{
compositeFromModelId(modelID = "kjzl6hvfrbw6c5yi3kx386312dpll69hi87xjlhm2t7m6cn7p1y4ojld1e9srgp", did_private_key= acc)
}
