% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDBgenerateAcc.R
\name{CDBgenerateAcc}
\alias{CDBgenerateAcc}
\title{CDBgenerateAcc}
\usage{
CDBgenerateAcc(key = NULL)
}
\arguments{
\item{key}{Character . key generated by CDBgenerateKey()}
}
\value{
Character . A DID repesenting the account
}
\description{
CDBgenerateAcc
}
\examples{
acc <- CDBgenerateAcc(key)
}
