% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDBModelCreate.R
\name{CDBModelCreate}
\alias{CDBModelCreate}
\title{CDBModelCreate}
\usage{
CDBModelCreate(
  content = NULL,
  ceramic_url = "http://localhost:7007",
  did_private_key = NULL
)
}
\arguments{
\item{content}{!!!!!!!! Figure uit hoe die model gemaak word - vanuit GraphQL ????}

\item{ceramic_url}{Character. Ceramic URL}

\item{did_private_key}{Character. Private key that will control model ??}
}
\description{
Most of the time you shouldn't be using this command directly.
Instead, you should first check if a model you need already exists, using Composites Discovery
and only if you can't find a model that you need, you should create one indirectly by
creating a Composite from a GraphQL Composite Schema.
}
\details{
CDBModelCreate
}
\references{
https://composedb.js.org/docs/0.3.x/api/commands/cli.model
}
