% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDBModelList.R
\name{CDBModelList}
\alias{CDBModelList}
\title{CDBModelList}
\usage{
CDBModelList(tab = TRUE, remoteURL = "http://46.101.111.94/", port = "7007")
}
\arguments{
\item{tab}{Logical}

\item{remoteURL}{Character or NULL}
}
\value{
Charachter
}
\description{
CDBModelList
}
\examples{
CDBModelList()
}
