% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_columnwise_tiles}
\alias{create_columnwise_tiles}
\title{Create columnwise tiles and return these as a list}
\usage{
create_columnwise_tiles(tiling, rowset, colgroups)
}
\arguments{
\item{tiling}{A tiling}

\item{rowset}{The rowset of the columnwise tiles}

\item{colgroups}{The column groups of the columnwise tiles: all columns in one group belong to the same tile}
}
\value{
A list of tiles
}
\description{
Create columnwise tiles and return these as a list
}
