% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_service}
\alias{create_service}
\title{Create a service.}
\usage{
create_service(data, path, dir, options = plotoptions())
}
\arguments{
\item{data}{an sf object}

\item{path}{service end-point name}

\item{dir}{service directory}
}
\description{
Create a service.
}
