% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_tile}
\alias{create_tile}
\title{Create a tile}
\usage{
create_tile(rowset, colset, dims)
}
\arguments{
\item{rowset}{Set of rows in the tile}

\item{colset}{Set of columns in the tile}

\item{dims}{The dimensions of the data matrix}
}
\value{
A tiling
}
\description{
Create a tile
}
