% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createifnot}
\alias{createifnot}
\title{create directory if it doesn't exist}
\usage{
createifnot(d)
}
\arguments{
\item{d}{directory name}
}
\value{
boolean
}
\description{
create directory if it doesn't exist
}
