% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCeramic.R
\name{createStream}
\alias{createStream}
\title{createStream}
\usage{
createStream(
  did,
  baseurl = "http://localhost:7007/",
  api_version = "v0",
  type = 0,
  family = "test2"
)
}
\arguments{
\item{did}{Character}

\item{baseurl}{Character. Default: "http://localhost:7007/"}

\item{api_version}{Character. Default "v0"}

\item{type}{Numeric. Default: 0}

\item{family}{Character. !!!  FIgure out what this is}
}
\description{
createStream
}
