% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df2mat}
\alias{df2mat}
\title{Convert a data frame to a matrix, returning the original column classes}
\usage{
df2mat(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A list containing the data frame as a matrix and the column classes
}
\description{
Convert a data frame to a matrix, returning the original column classes
}
