% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dist_maximum_change_in_correlation}
\alias{dist_maximum_change_in_correlation}
\title{Given two datasets ds1 and ds2, find the best variable pair for which  the difference in correlation between ds1 and ds2 is maximum}
\usage{
dist_maximum_change_in_correlation(ds1, ds2, data = FALSE)
}
\arguments{
\item{ds1}{A dataset (data frame or matrix)}

\item{ds2}{A dataset (data frame or matrix)}
}
\value{
List containing (i) correlation values for the pairs, (ii) the correlation matrices, (iii) the pairs of matrices with the largest change in correlation, (iv) the names of the ordered pairs of columns
}
\description{
Given two datasets ds1 and ds2, find the best variable pair for which  the difference in correlation between ds1 and ds2 is maximum
}
