% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{enum_slices}
\alias{enum_slices}
\title{Determine in how many pieces a new tile will be divided when added to an existing tiling}
\usage{
enum_slices(tiling, tile)
}
\arguments{
\item{tiling}{A tiling}

\item{tile}{A new tile to be added to the tiling}
}
\value{
A matrix containing for each row in the data matrix affected by the new tile:
          (i) a "slice number" enumerating the slices
         (ii) the indices of the rows in the data matrix affected by the new tile
        (iii) the existing permutations affected by the addition of the new tile
}
\description{
When a new tile is added to an existing tiling, it must be broken up in horizontal slices
if it overlaps with existing tiles in the tiling. A slice indicates a set of overlapping
permutations that must be merged. This function determines in how many slices the new tile
will be broken up and enumerates these slices. One slice spans one or more rows, and
the same permutation is applied to all elements within a slice.
}
