% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_best_variable_pair}
\alias{find_best_variable_pair}
\title{Given two datasets ds1 and ds2, find the best variable pair for which  the difference in correlation between ds1 and ds2 is maximum}
\usage{
find_best_variable_pair(ds1, ds2, method = "average_correlation", ...)
}
\arguments{
\item{ds1}{A dataset (data frame or matrix)}

\item{ds2}{A dataset (data frame or matrix)}

\item{method}{Which method should be used (currenlty only \code{max_change_in_correlation} supported.}

\item{...}{Extra arguments to be passed to the method used.}
}
\value{
Depends on the used method.
}
\description{
Given two datasets ds1 and ds2, find the best variable pair for which  the difference in correlation between ds1 and ds2 is maximum
}
