% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCeramicAccFromProfile.R
\name{getCeramicAccFromProfile}
\alias{getCeramicAccFromProfile}
\title{getCeramicAccFromProfile}
\usage{
getCeramicAccFromProfile(file = "~/.ceramic/daemon.config.json")
}
\arguments{
\item{file}{Character. Location of daemon.config.json. Default ~/.ceramic/daemon.config.json}
}
\value{
Character. DID
}
\description{
getCeramicAccFromProfile
}
\examples{
acc1 <- getCeramicAccFromProfile()
}
