% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCeramicProfile.R
\name{getCeramicProfile}
\alias{getCeramicProfile}
\title{getCeramicProfile}
\usage{
getCeramicProfile(file = "~/.ceramic/daemon.config.json")
}
\arguments{
\item{file}{Character. Location of daemon.config.json. Default ~/.ceramic/daemon.config.json}
}
\value{
list
}
\description{
getCeramicProfile
}
\examples{
profile <- getCeramicProfile(file = "~/.ceramic/daemon.config.json")
}
