% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCeramicStatus.R
\name{getCeramicStatus}
\alias{getCeramicStatus}
\title{getCeramicStatus}
\usage{
getCeramicStatus(key, url = "http://localhost:7007/")
}
\arguments{
\item{key}{Character.}

\item{url}{Character.}
}
\description{
getCeramicStatus
}
\examples{
getCeramicStatus()
}
