% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCeramicStream.R
\name{getCeramicStream}
\alias{getCeramicStream}
\title{getCeramicStream}
\usage{
getCeramicStream(URL = CERAMIC_URL, SID = mtcarSID, contentAsTibble = FALSE)
}
\arguments{
\item{URL}{Ceramic URL . Default CERAMIC_URL http://146.190.193.80:7007}

\item{SID}{StreamID like "kjzl6cwe1jw146ct4uzdlismretwt0o5alh3brpelfc49klnq3xwd610r61mrox"}

\item{contentAsTibble}{Logical. If TRUE return content - seems to be equivalent to glaze_tile_content()}
}
\description{
getCeramicStream
}
\examples{
getCeramicStream(URL = "http://146.190.6.248:7007", SID = "kjzl6cwe1jw14blkfyymlen4d48ljm1dxl9tijy975b6n8v62m1uwbiy1lz00vn", contentAsTibble = TRUE)
}
