% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComposedbModelContent.R
\name{getComposeDBModelContent}
\alias{getComposeDBModelContent}
\title{getComposeDBModelContent}
\usage{
getComposeDBModelContent(
  url = "http://localhost:7007",
  apiaddres = "/api/v0/streams/",
  modelid = "kjzl6hvfrbw6c8pwn8vkbueq611nkptorifzqhog6e5b4z01fzuynuw00edzx4v"
)
}
\arguments{
\item{url}{Character. Ceramic URI. Default "http://localhost:7007"}

\item{apiaddres}{Character. Path to api endpoint. Default "/api/v0/streams/"}

\item{modelid}{Character.}
}
\value{
tibble
}
\description{
getComposeDBModelContent
}
\examples{
getComposeDBModelContent(url = "http://146.190.6.248:7007", modelid = "kjzl6hvfrbw6c8pwn8vkbueq611nkptorifzqhog6e5b4z01fzuynuw00edzx4v" )
}
