% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getK8sSecrets.R
\name{getK8sSecrets}
\alias{getK8sSecrets}
\title{getK8sSecrets}
\usage{
getK8sSecrets(
  namespace = "ceramic",
  name = "ceramic-admin",
  field = "data",
  subfield = "private-key"
)
}
\arguments{
\item{namespace}{Character}

\item{name}{Character. Default: "ceramic-admin" . Use listK8sSecrets()$NAME to see all}

\item{field}{Character. Default "data"}

\item{subfield}{Character. Default "private-key"}
}
\value{
Character
}
\description{
getK8sSecrets
}
\examples{
getK8sSecrets()
}
