% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getK8sServices.R
\name{getK8sServices}
\alias{getK8sServices}
\title{getK8sServices}
\usage{
getK8sServices(
  namespace = "ceramic",
  type = c("NodePort", "LoadBalancer", "ClusterIP")[2]
)
}
\arguments{
\item{namespace}{Character. K8s namespace. Use listK8sNamespaces()$NAME to see options}

\item{type}{Character. Service type. Posibly: "NodePort", "LoadBalancer", "ClusterIP"}
}
\value{
tibble with "NAME", "TYPE", "CLUSTER_IP", "EXTERNAL_IP", "PORTS", "AGE"
}
\description{
getK8sServices
}
\examples{
getK8sServices()$NAME
getK8sServices()$EXTERNAL_IP
getK8sServices(type = NULL) # Alles
}
