% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glaze_tile.R
\name{glaze_tile_content}
\alias{glaze_tile_content}
\title{glaze_tile_content}
\usage{
glaze_tile_content(STREAMID, CERAMIC_URL = NULL, k = NULL, returnJSON = FALSE)
}
\arguments{
\item{STREAMID}{Character.}

\item{CERAMIC_URL}{Character.}

\item{k}{Character.}

\item{returnJSON}{Logical. Return JSON or not}
}
\value{
tibble
}
\description{
glaze_tile_content
}
