% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glaze_tile.R
\name{glaze_tile_create}
\alias{glaze_tile_create}
\title{glaze_tile_create}
\usage{
glaze_tile_create(CERAMIC_URL = NULL, k = NULL, content = NULL)
}
\arguments{
\item{k}{Character. Key}

\item{content}{Character. JSON with data}

\item{URL}{Character. Ceramic URL}
}
\description{
glaze_tile_create
}
\examples{
k <- "451c064691fe121292d2eaa05f552de8aeb2e4136c03c546c209f89e96263510"
een <- jsonlite::fromJSON("Data/Een.json")
twee <- '{"Foo":"Bar"}'
tweeSID <- glaze_tile_create(k = k, content = twee)
mtcarSID <- glaze_tile_create(k = k, content = jsonlite::toJSON(mtcars[1:2,]))
}
