% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glaze_tile_update.R
\name{glaze_tile_update}
\alias{glaze_tile_update}
\title{glaze_tile_update}
\usage{
glaze_tile_update(
  STREAMID,
  CERAMIC_URL = NULL,
  k = NULL,
  m = NULL,
  content = NULL
)
}
\arguments{
\item{STREAMID}{Character.}

\item{CERAMIC_URL}{Character.}

\item{k}{Character.}

\item{m}{Character.}

\item{content}{Character. JSON of some kind}
}
\value{
Character. streamID
}
\description{
glaze_tile_update
}
\examples{
glaze_tile_update(STREAMID = mtcarSID, CERAMIC_URL = REMOTE_URL, k = DID_PRIVATE_KEY, content = jsonlite::toJSON(mtcars[3:4,]))
}
