% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ind2sub}
\alias{ind2sub}
\title{Linear index to subscript}
\usage{
ind2sub(ind, nrows)
}
\arguments{
\item{ind}{Linear index}

\item{nrows}{The number of rows in the data matrix}
}
\value{
The subscript index (row, col) corresponding to the linear index ind
}
\description{
Linear index to subscript
}
