% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{init_tiles}
\alias{init_tiles}
\title{Initialise tiles in a tiling}
\usage{
init_tiles(dims, columnwise_tiles = FALSE)
}
\arguments{
\item{dims}{Dimensions of the data matrix}

\item{columnwise_tiles}{Boolean indicating if the tiling should be initialised with columnwise tiles or not. Default is FALSE.}
}
\value{
A matrix used to describe tiles
}
\description{
Initialise tiles in a tiling
}
