% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{init_tiling}
\alias{init_tiling}
\title{Initialise a tiling}
\usage{
init_tiling(dims, columnwise_tiles = FALSE)
}
\arguments{
\item{dims}{Dimensions of the data matrix}

\item{columnwise_tiles}{Boolean indicating if the tiling should be initialised with columnwise tiles or not. Default is FALSE.}
}
\value{
A tiling, permuting using which leaves everything intact.
}
\description{
Initialise a tiling
}
