% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listK8sSecrets.R
\name{listK8sSecrets}
\alias{listK8sSecrets}
\title{listK8sSecrets}
\usage{
listK8sSecrets(namespace = "ceramic")
}
\arguments{
\item{namespace}{Character. to see all use: system("kubectl get namespaces")}
}
\description{
listK8sSecrets
}
