% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lnglat}
\alias{lnglat}
\title{Convert xy to lnglat}
\usage{
lnglat(x, y, zoom)
}
\arguments{
\item{x}{x-position on grid}

\item{y}{y-position on grid}

\item{zoom}{zoom level}
}
\value{
lng/lat representing top left corner of slippy-map tile
}
\description{
Convert xy to lnglat
}
