% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_random_tile}
\alias{make_random_tile}
\title{Make a random tile}
\usage{
make_random_tile(nr, nc)
}
\arguments{
\item{nr}{The number of rows in the data matrix}

\item{nc}{The number of columns in the data matrix}
}
\value{
A tile
}
\description{
Make a random tile
}
