% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGenesisCommit.R
\name{makeGenesisCommit}
\alias{makeGenesisCommit}
\title{makeGenesisCommit}
\usage{
makeGenesisCommit(
  type = "0",
  family = "test",
  controllers = c("did:key:z6MkfZ6S4NVVTEuts8o5xFzRMR8eC6Y1bngoBQNnXiCvhH8H"),
  returnJSON = TRUE
)
}
\arguments{
\item{type}{Character. see MakeStreamtype(). Remember to use as.integer(code, base = 16)}

\item{family}{Character.}

\item{controllers}{Character. DID}

\item{returnJSON}{Logical. Return JSON (default) or R list}
}
\value{
list or if returnJSON is TRUE, JSON
}
\description{
Create a Genesis commit
}
