% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSteamtypes.R
\name{makeNetworkIds}
\alias{makeNetworkIds}
\title{makeNetworkIds}
\usage{
makeNetworkIds()
}
\description{
make a tibble with Network IDs
}
\examples{
dfIDs <- makeNetworkIds()
}
\references{
https://github.com/ceramicnetwork/CIPs/tree/main/tables
}
