% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSteamtypes.R
\name{makeStreamtype}
\alias{makeStreamtype}
\title{MakeStreamtype}
\usage{
makeStreamtype()
}
\description{
Convenience function to creat table of streamtypes
}
\examples{
dfTypes <- makeStreamtype()
}
\references{
https://github.com/ceramicnetwork/CIPs/tree/main/tables
}
