% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mat2df}
\alias{mat2df}
\title{Convert a matrix to a data frame, also setting column classes}
\usage{
mat2df(x, col_classes)
}
\arguments{
\item{x}{A data frame}

\item{col_classes}{A list containing the column class for each column in the matrix / data frame}
}
\value{
A data frame
}
\description{
Convert a matrix to a data frame, also setting column classes
}
