% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{metadata}
\alias{metadata}
\alias{tile_meta}
\alias{tile_origin}
\alias{tile_bounds}
\alias{tile_res}
\title{Tile metadata}
\usage{
tile_meta(tiles)

tile_origin(tiles)

tile_bounds(tiles)

tile_res(tiles)
}
\arguments{
\item{tiles}{character, tiles directory.}
}
\value{
a list or vector.
}
\description{
Get tile metadata from tile set.
}
\details{
These functions return metadata entries from the \code{tilemapresource.xml} file located in the \code{tiles} directory. Any tiles generated with \code{tiler} include this file.
\code{tile_meta} returns a list containing the origin (y, x), bounding box (ymin, xmin, ymax, xmax), and zoom resolutions (zoom order low to high), formatted ready for use with \code{leaflet::leafletCRS}.
The other wrapper functions return the respective list element.
}
\examples{
\dontshow{tmpfiles <- list.files(tempdir(), full.names = TRUE)}
x <- system.file("maps/map_wgs84.tif", package = "tiler")
tiles <- file.path(tempdir(), "tiles")
tile(x, tiles, "2-3")
tile_meta(tiles)
tile_origin(tiles)
tile_bounds(tiles)
tile_res(tiles)
\dontshow{
unlink(c(tiles, file.path(tempdir(), "preview.html")), recursive = TRUE, force = TRUE)
extrafiles <- setdiff(list.files(tempdir(), full.names = TRUE), tmpfiles)
if(length(extrafiles)) unlink(extrafiles, recursive = TRUE, force = TRUE)
}
}
