% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{permute_data}
\alias{permute_data}
\title{Permute a data matrix using a tiling}
\usage{
permute_data(x, tiling, rseed = NULL)
}
\arguments{
\item{x}{A data matrix or a data frame}

\item{tiling}{A tiling}

\item{rseed}{Random seed used for permuting.}
}
\value{
The data matrix permuted according to the tiling
}
\description{
Permute a data matrix using a tiling
}
