% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{permute_vector}
\alias{permute_vector}
\title{Permute a vector}
\usage{
permute_vector(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The vector in which the elements have been shuffled
}
\description{
Permute a vector
}
