% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pixelsize}
\alias{pixelsize}
\title{get pixel size in degrees}
\usage{
pixelsize(bb)
}
\arguments{
\item{bb}{bounding box created by `bbox` function}
}
\value{
vector with width and height of a pixel in degrees
}
\description{
get pixel size in degrees
}
