% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotoptions}
\alias{plotoptions}
\title{Plotting options}
\usage{
plotoptions(...)
}
\arguments{
\item{pal}{function which returns fill color}

\item{value}{column holding value for pal}

\item{col}{fill color, not used if `pal` and `value` provided}

\item{border}{border color}

\item{lwd}{line width}

\item{lty}{line type}
}
\description{
Plotting options
}
