% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print_tiling}
\alias{print_tiling}
\title{Helper function for printing out a tiling}
\usage{
print_tiling(tiling)
}
\arguments{
\item{tiling}{A tiling containing user and base tile. If the tiling has columnwise tiles set as an option, these are hounoured.}
}
\value{
Nothing
}
\description{
This function prints the permutations in a tiling as a matrix
}
