% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{printm}
\alias{printm}
\title{Helper function for printing out a tiling}
\usage{
printm(tiling, col)
}
\arguments{
\item{tiling}{A tiling containing user and base tile. If the tiling has columnwise tiles set as an option, these are hounoured.}
}
\value{
Nothing
}
\description{
This function prints out a specific column from a tiling.
}
