% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rebuild_tiling}
\alias{rebuild_tiling}
\title{Rebuild tiling.}
\usage{
rebuild_tiling(tiling)
}
\arguments{
\item{tiling}{A tiling containing user and base tile. If the tiling has columnwise tiles set as an option, these are hounoured.}
}
\value{
A tiling
}
\description{
This function rebuilds a tiling from scratch, using both the user and base tiles
}
