% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_tiles}
\alias{remove_tiles}
\title{Remove a given tile from the tiling}
\usage{
remove_tiles(tiling, tile_ind, target)
}
\arguments{
\item{tiling}{A tiling. If the tiling already has columnwise tiles set as an option, these are hounoured.}

\item{tile_ind}{Index of the tile(s) to be removed from the user tile list}

\item{target}{List from which the tile should be removed. One of \code{tile_list_user}, \code{tile_list_base} or \code{tile_list_ignore}}
}
\value{
The tiling without the given tile
}
\description{
Remove a given tile from the tiling
}
