% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCeramic.R
\name{runCeramic}
\alias{runCeramic}
\title{runCeramic}
\usage{
runCeramic(network = "testnet-clay")
}
\arguments{
\item{network}{Character. Network name}
}
\value{
Runs Ceramic deamon as side effect
}
\description{
runCeramic
}
\examples{
runCeramic()
}
