% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupComposedbK8s.R
\name{setupComposedbK8s}
\alias{setupComposedbK8s}
\title{setupComposedbK8s
Set up connection to kubernetes cluster running composedb.
Activates port forwarding}
\usage{
setupComposedbK8s(
  namespace = "ceramic",
  localport = "7007",
  k8sport = "7007",
  secretname = "ceramic-admin",
  DID = NULL,
  pk = NULL,
  URL = NULL,
  returnlist = FALSE
)
}
\arguments{
\item{namespace}{Character. to see all use: Tiler::listK8sNamespaces() or system("kubectl get namespaces")}

\item{localport}{Character or numeric . local port used by composedb cli}

\item{k8sport}{Character or numeric . remote port used by composedb}

\item{secretname}{Character. Default "ceramic-admin"}

\item{DID}{Character}

\item{pk}{Character}

\item{URL}{Character}

\item{returnlist}{Logical. If FALSE (default), returns the CERAMIC_URL. If TRUE returns a list with DID_PRIVATE_KEY and CERAMIC_URL}
}
\description{
setupComposedbK8s
Set up connection to kubernetes cluster running composedb.
Activates port forwarding
}
\examples{
setupComposedbK8s()
}
