% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{smax}
\alias{smax}
\title{Special max function}
\usage{
smax(x, default = 1)
}
\arguments{
\item{x}{A vector}

\item{default}{Value to return if the input vector is empty or if max would return -Inf/+Inf.}
}
\value{
The vector in which the elements have been shuffled
}
\description{
This function ignores NA values and returns a default value if the input vector is empty or if max would return -Inf/+Inf.
}
