% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sub2ind_set}
\alias{sub2ind_set}
\title{Row and column sets to linear indexing}
\usage{
sub2ind_set(rowset, colset, dims)
}
\arguments{
\item{rowset}{Set of rows}

\item{colset}{Set of columns}

\item{dims}{Dimensions of the data matrix}
}
\value{
The (column-major) linear indices corresponding to the row and column sets
}
\description{
Row and column sets to linear indexing
}
