% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiler.R
\name{tile}
\alias{tile}
\title{Create map tiles from an image}
\usage{
tile(file, tile_dir, zoom, crs = NULL, ...)
}
\arguments{
\item{file}{character, image file.}

\item{tile_dir}{character, output directory for generated tiles.}

\item{zoom}{character, zoom levels. Example format" \code{"3-7"}.}

\item{crs}{character, proj4 string. Use this to force set the CRS of a loaded raster object from \code{file} in cases where the CRS is missing but known, to avoid defaulting to non-geographic tiling.}

\item{...}{additional arguments for projected maps: reprojection method or any arguments to \code{raster::RGB}, e.g. \code{col} and \code{colNA}. See details.}
}
\value{
nothing is returned but tiles are written to disk.
}
\description{
Create geographic and non-geographic map tiles from an image file.
}
\details{
This function supports both geographic and non-geographic tile generation.
When \code{file} is a simple image file such as \code{png}, \code{tile} generates non-geographic, simple CRS tiles.
Files that can be loaded by the \code{raster} package yield geographic tiles, as long as \code{file} has projection information.
If the raster object's proj4 string is \code{NA}, it falls back on non-geographic tile generation and a warning is thrown.

Supported simple CRS/non-geographic image file types include \code{png}, \code{jpg} and \code{bmp}.
For projected map data, supported file types include three types readable by the \code{raster} package: \code{grd}, \code{tif}, and \code{nc} (requires \code{ncdf4}).
Other currently unsupported file types passed to \code{file} throw an error.

If a map file loadable by \code{raster} is a single-layer raster object, tile coloring is applied.
To override default coloring of data and \code{noData} pixels, pass the additional arguments \code{col} and \code{colNA} to \code{...}.
Multi-layer raster objects are rejected with an error message. The only exception is a three- or four-band raster, assumed to represent red, green, blue and alpha channels, respectively.
In this case, processing will continue but coloring is ignored as unnecessary.

Prior to tiling, a geographically-projected raster layer is reprojected to have the proj4 string: \code{+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs}, only if it has some other projection.
Otherwise no reprojection is needed.
The only reprojection argument available through \code{...} is \code{method}, which can be \code{"bilinear"} (default) or\code{"ngb"}.
If complete control over reprojection is required, this should be done prior to passing the rasterized file to the \code{tile} function. Then no reprojection is performed by \code{tile}.

Working with the \code{leaflet} package. In contrast to non-geographic tiles, for geographic tiles you may need to set \code{tms: true} in your Leaflet code for tiles to arrange properly.
See the Leaflet JS library and \code{leaflet} package documentation for working with custom tiles in Leaflet.
}
\examples{
# non-geographic/simple CRS
x <- system.file("maps/map.png", package = "tiler")
tile(x, tempdir(), "0")

# projected map
\dontrun{
x <- system.file("maps/map.tif", package = "tiler")
tile(x, tempdir(), "0")
}
}
