% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{tiles}
\alias{tiles}
\alias{tiles.default}
\alias{tiles.RasterLayer}
\title{Tiles}
\usage{
tiles(x, ...)

\method{tiles}{default}(x, ...)

\method{tiles}{RasterLayer}(x, output_dir, colors = "Spectral",
  zoom = "5-16", threads = 1, project = TRUE, verbose = FALSE,
  cache = TRUE, shiny = FALSE, ...)
}
\arguments{
\item{x}{\code{\link[raster]{RasterLayer-class}} object.}

\item{...}{not used.}

\item{output_dir}{\code{\link{character}} directory in which to save tiles.}

\item{colors}{The color palette (see \code{\link{colorNumeric}}) or function
to use to color the raster values (hint: if providing a function, set
\code{na.color} to \code{"#00000000"} to make \code{NA} areas transparent).
Defaults to \code{"Spectral"}.}

\item{zoom}{\code{character} Zoom levels to render. The levels can be
specified as a single number (e.g \code{"10"}) or as a range
(e.g. \code{"1-10"}). Defaults to \code{"5-16"}.}

\item{threads}{\code{numeric} number of threads to use for processing.
Defaults to 1.}

\item{project}{\code{logical} If argument to \code{x} is not already in the
coordinate reference system expected by Leaflet (Mercator;
\code{EPSG:3857}), should the data be automatically reprojected? If
argument to \code{project} is \code{FALSE}, then the user will need
to ensure that the data is in the correct coordinate reference system.}

\item{verbose}{\code{logical} Should information be printed during
processing?}

\item{cache}{\code{logical} if tiles already exist in argument to
\code{output_dir} should the processing be skipped?}

\item{shiny}{\code{logical} are the tiles being served in a shiny
applications?}
}
\value{
\code{character} tile scheme.
}
\description{
Create tiles for a spatial dataset.
}
\details{
This function creates a tile layer, following the
  \href{https://wiki.osgeo.org/wiki/Tile_Map_Service_Specification}{Tile Map Service specification}, using the object in argument to \code{x} and saves it
  in the argument to \code{output_dir}.
}
\examples{
# load leafleat
library(leaflet)

# load example data set
data(netherlands, package = "mapmisc")

\donttest{
# create leaflet map and visualize data using tiles
map <- leaflet() \%>\%
       setView(lng = 5.75560, lat = 50.94723, zoom = 12) \%>\%
       addProviderTiles("Esri.WorldImagery", group = "Basemap") \%>\%
       addTiles(tiles(nldElev, output_dir = tempdir(), zoom = "10-16"),
                group = "Elevation",
                options = tileOptions(minZoom = 5, maxZoom = 16,
                                      tms = TRUE)) \%>\%
       addLayersControl(baseGroups = "Basemap", overlayGroups = "Elevation",
                        options = layersControlOptions(collapsed = FALSE))

# render map
print(map)
}
}
\seealso{
\code{\link[leaflet]{addTiles}}.
}
