% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCeramicConfig.R
\name{updateCeramicConfig}
\alias{updateCeramicConfig}
\title{updateCeramicConfig}
\usage{
updateCeramicConfig(
  file = "~/.ceramic/daemon.config.json",
  variable = c("anchor", "http-api", "ipfs", "logger", "metrics", "network", "node",
    "state-store", "indexing")[2],
  subvariable = NULL,
  value = NULL,
  outpath = NULL
)
}
\arguments{
\item{file}{Character  Path to daemon.config.json file. Default:  "~/.ceramic/daemon.config.json"}

\item{variable}{Character . One of variable = c("anchor", "http-api", "ipfs", "logger",
"metrics", "network", "node", "state-store", "indexing"). Default "http-api"}

\item{subvariable}{Character. Name of subvariable to change or NULL. One off:
list(anchor = "auth-method",
    http-api = c("cors-allowed-origins", "admin-dids"),
    ipfs = "mode", logger = c("log-level", "log-to-files"),
    metrics = "metrics-exporter-enabled",
    network = "name",
    node = "privateSeedUrl",
    state-store = c("mode", "local-directory"),
    indexing = c("db", "allow-queries-before-historical-sync", "disable-composedb", "enable-historical-sync"))}

\item{outpath}{Character. Filepath to write output to. If NULL file is written to original location (file)}

\item{value.}{Character. Value to set variable[[subvariable]] to.}
}
\value{
json
}
\description{
updateCeramicConfig
}
\examples{
updateCeramicConfig(variable = "http-api", subvariable = "admin-dids", value = acc)
}
