% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xy}
\alias{xy}
\title{Convert lng/lat to xy grid position}
\usage{
xy(lng, lat, zoom)
}
\arguments{
\item{lng}{longitude}

\item{lat}{latitude}

\item{zoom}{zoom level}
}
\value{
xy representing top-left corner of a slippy-map tile.
}
\description{
Convert lng/lat to xy grid position
}
