% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_tile}
\alias{make_tile}
\title{create tile image}
\usage{
make_tile(data, x, y, zoom, path, options = plotoptions(), cache = tempdir())
}
\arguments{
\item{data}{an sf object}

\item{x}{x-grid}

\item{y}{y-grid}

\item{zoom}{zoom level}

\item{path}{name of tileset folder}

\item{cache}{directory for cache}

\item{style}{Not implemented.  Intended to support basic styling.}
}
\value{
path to tile
}
\description{
create tile image
}
