% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tiler}
\alias{tiler}
\alias{tiler-package}
\title{tiler}
\description{
The \emph{tiler} package is used to create tile maps to render large
spatial data sets.
}
\details{
This package is designed to be used with the \emph{leaflet}
package. The main function in this package is \code{\link{create_tiles}} and
can prepare tile data for use in \code{\link[leaflet]{addWMSTiles}}.
}
