% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiler.R
\docType{package}
\name{tiler}
\alias{tiler}
\alias{tiler-package}
\title{tiler: Create map tiles from images.}
\description{
The tiler package creates map tiles for geographic and non-geographic ("simple CRS") images.
}
