% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiler.R
\docType{package}
\name{tiler}
\alias{tiler}
\alias{tiler-package}
\title{tiler: Create map tiles from images.}
\description{
The tiler package creates map tiles for geographic and non-geographic ("simple CRS") images.
}
\details{
This package provides a map tile-generator function for creating map tile sets for use with packages such as \code{leaflet}.
In addition to generating map tiles based on a common raster layer source, it also handles the non-geographic edge case, producing map tiles from arbitrary images.
These map tiles, which have a 'simple CRS', a non-geographic simple Cartesian coordinate reference system, can also be used with \code{leaflet} when applying the simple CRS option.
\cr\cr
Map tiles can be created from a input file with any of the following extensions: 'tif', 'grd' and 'nc' for spatial maps and 'png', 'jpg' and 'bmp' for basic images.
\cr\cr
This package requires Python and the \code{gdal} library for Python.
Windows users are recommended to install \href{OSGeo4W}{https://trac.osgeo.org/osgeo4w/} as an easy way to obtain the required \code{gdal} support for Python in Windows.
}
